fix - questions not updated on websocket response
